refactor(db): split table definitions by domain
